refactor: invert activity report's agent-task source behind a hook
core::activity listed durable agent-task records (agent_task_lifecycle::
list_records / record_health_summary), projected each into an ActivityItem,
and stored the AgentTaskRecordHealthSummary on the report — a core->agent-
task edge in a core file that stays in core.

Invert it behind an ActivityAgentTaskProvider hook: core owns the activity
report and the ActivityItem shape; the agent-task layer supplies the items
and the health summary. The item-projection logic (item_from_agent_task /
actions_for_agent_task) moves to the agent-side provider (which builds core's
pub ActivityItem). The report's agent_task_record_health field becomes an
opaque serde_json::Value. With no provider registered the no-op contributes
no items and a null health summary.

Stage 3 prep for the homeboy-agents extraction: severs activity.rs's
agent-task dependency so core stays agent-type-free before the git-mv.
