+++ crates/homeboy-cli/src/cli_runtime.rs
@@ -181,6 +181,10 @@ impl CliRuntime {
         // agent-task gate-feedback candidate without depending on the agent-task
         // subsystem.
         crate::core::agent_task_candidate_baseline::register();
+        // Register the agent-task activity provider so core's activity report
+        // includes durable agent-task records and their health summary without
+        // depending on the agent-task subsystem.
+        crate::core::agent_task_lifecycle::activity_provider::register();
         // Register the command-label resolver so core::runner can map dispatched
         // argv to a hot-command label without depending on the full CLI parser.
         crate::runner::set_command_label_resolver(|argv| {
