+++ crates/homeboy-core/src/agent_task_lifecycle/mod.rs
@@ -29,6 +29,7 @@ mod lifecycle_store;
 
 use lifecycle_store as store;
 
+pub mod activity_provider;
 pub mod agent_task_lifecycle_event;
 mod artifact_materialization;
 mod cancellation;
